Refactor PR comments, move logic to new service
